+++ src/core/Table/Table.tsx
@@ -118,6 +118,11 @@ export type TableProps<
    * Content shown when there is no data after filtering.
    */
   emptyFilteredTableContent?: React.ReactNode;
+  /**
+   * Function that should return true if a row is disabled (i.e. cannot be selected or expanded).
+   * If not specified, all rows are enabled.
+   */
+  isRowDisabled?: (rowData: T) => boolean;
 } & Omit<CommonProps, 'title'>;
 
 /**
@@ -188,6 +193,7 @@ export const Table = <
     emptyFilteredTableContent,
     filterTypes: filterFunctions,
     expanderCell,
+    isRowDisabled = () => false,
     ...rest
   } = props;
 
@@ -242,6 +248,7 @@ export const Table = <
                 onClick={() => {
                   row.toggleRowExpanded();
                 }}
+                disabled={isRowDisabled(props.row.original)}
               >
                 {<SvgChevronRight />}
               </IconButton>
@@ -267,11 +274,34 @@ export const Table = <
         maxWidth: 48,
         columnClassName: 'iui-slot',
         cellClassName: 'iui-slot',
-        Header: ({ getToggleAllRowsSelectedProps }: HeaderProps<T>) => (
-          <Checkbox {...getToggleAllRowsSelectedProps()} />
-        ),
+        Header: ({ getToggleAllRowsSelectedProps }: HeaderProps<T>) => {
+          const disabled = instance.rows.every((row) =>
+            isRowDisabled(row.original),
+          );
+          const checked =
+            instance.isAllRowsSelected ||
+            instance.rows.every(
+              (row) =>
+                instance.state.selectedRowIds[row.id] ||
+                isRowDisabled(row.original),
+            );
+          return (
+            <Checkbox
+              {...getToggleAllRowsSelectedProps()}
+              checked={checked && !disabled}
+              indeterminate={
+                !checked &&
+                Object.keys(instance.state.selectedRowIds).length > 0
+              }
+              disabled={disabled}
+            />
+          );
+        },
         Cell: ({ row }: CellProps<T>) => (
-          <Checkbox {...row.getToggleRowSelectedProps()} />
+          <Checkbox
+            {...row.getToggleRowSelectedProps()}
+            disabled={isRowDisabled(row.original)}
+          />
         ),
       },
       ...columns,
@@ -317,11 +347,17 @@ export const Table = <
     }
 
     const selectedData: T[] = [];
-    instance.rows.forEach((row) => {
-      if (newState.selectedRowIds[row.id]) {
-        selectedData.push(row.original);
+    const newSelectedRowIds = {} as Record<string, boolean>;
+    Object.keys(newState.selectedRowIds).forEach((id) => {
+      if (
+        newState.selectedRowIds[id] &&
+        !isRowDisabled(instance.rowsById[id].original)
+      ) {
+        newSelectedRowIds[id] = true;
+        selectedData.push(instance.rowsById[id].original);
       }
     });
+    newState.selectedRowIds = newSelectedRowIds;
     onSelect?.(selectedData, newState);
   };
 
@@ -483,6 +519,7 @@ export const Table = <
               className: cx('iui-row', {
                 'iui-selected': row.isSelected,
                 'iui-row-expanded': row.isExpanded && subComponent,
+                'iui-disabled': isRowDisabled(row.original),
               }),
             });
             return (